use unsigned long

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -9,7 +9,7 @@
 
 int main(void)
 {
-long long i, n = 612852475143;
+unsigned long i, n = 612852475143;
 
 for (i = 2; i <= sqrt(n); i++)
 {
